Add comments

+++ ContosoUniversity/Controllers/StudentsController.cs
@@ -19,14 +19,17 @@ namespace ContosoUniversity.Controllers
             _context = context;
         }
 
-        // GET: Students
+        // GET: Students | with SortOrder paramater getting the query string action method paramater from URL. 
+        // All params will be null if page displayed for first time or hasn't clicked a paging or sorting link.
         public async Task<IActionResult> Index(
      string sortOrder,
      string currentFilter,
      string searchString,
      int? pageNumber)
         {
+            //Provides the view with the current sort order.
             ViewData["CurrentSort"] = sortOrder;
+            //If sortOrder param is null/empty, set NameSort param as descending. 
             ViewData["NameSortParm"] = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
             ViewData["DateSortParm"] = sortOrder == "Date" ? "date_desc" : "Date";
 
@@ -48,6 +51,7 @@ namespace ContosoUniversity.Controllers
                 students = students.Where(s => s.LastName.Contains(searchString)
                                        || s.FirstMidName.Contains(searchString));
             }
+            //Ascending order 
             switch (sortOrder)
             {
                 case "name_desc":
@@ -59,6 +63,7 @@ namespace ContosoUniversity.Controllers
                 case "date_desc":
                     students = students.OrderByDescending(s => s.EnrollmentDate);
                     break;
+                //Default sortOrder is ascending
                 default:
                     students = students.OrderBy(s => s.LastName);
                     break;
@@ -154,7 +159,8 @@ namespace ContosoUniversity.Controllers
 		    var studentToUpdate = await _context.Students.FirstOrDefaultAsync(s => s.ID == id);
 		    if (await TryUpdateModelAsync<Student>(
 		        studentToUpdate,
-		        "",
+                //Empty string is a prefix to use with the form fields names.
+                "",
 		        s => s.FirstMidName, s => s.LastName, s => s.EnrollmentDate))
 		    {
 		        try
@@ -174,7 +180,7 @@ namespace ContosoUniversity.Controllers
 		}
 
         // GET: Students/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -182,10 +188,18 @@ namespace ContosoUniversity.Controllers
             }
 
             var student = await _context.Students
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (student == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
             }
 
             return View(student);
@@ -197,9 +211,22 @@ namespace ContosoUniversity.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var student = await _context.Students.FindAsync(id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (student == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Students.Remove(student);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool StudentExists(int id)
